fix even more warnings reported by phpstan

+++ classes/counters.php
@@ -120,7 +120,8 @@ class Counters {
 				$has_img = false;
 			}
 
-			if (date('Y') - date('Y', strtotime($line['last_updated'])) > 2)
+			// hide default un-updated timestamp i.e. 1980-01-01 (?) -fox
+			if ((int)date('Y') - (int)date('Y', strtotime($line['last_updated'])) > 2)
 				$last_updated = '';
 
 			$cv = [
